TD/reponses/sources/automatisation/requirements.txt
requests
configparser
google-cloud-pubsub
google-cloud-storage
